CBooRenderer: Make use of std::array where applicable
Makes the data strongly typed and allows eliminating some hardcoded
sizes.

+++ Runtime/Graphics/CBooRenderer.hpp
@@ -176,10 +176,10 @@ class CBooRenderer final : public IRenderer {
   void ActivateLightsForModel(CAreaListItem* item, CBooModel& model);
   void RenderBucketItems(CAreaListItem* item);
   void HandleUnsortedModel(CAreaListItem* item, CBooModel& model, const CModelFlags& flags);
-  static void CalcDrawFogFan(const zeus::CPlane* planes, int numPlanes, const zeus::CVector3f* verts, int numVerts,
-                             int iteration, int level, CFogVolumePlaneShader& fogVol);
-  static void DrawFogSlices(const zeus::CPlane* planes, int numPlanes, int iteration, const zeus::CVector3f& center,
-                            float delta, CFogVolumePlaneShader& fogVol);
+  static void CalcDrawFogFan(const zeus::CPlane* planes, size_t numPlanes, const zeus::CVector3f* verts,
+                             size_t numVerts, size_t iteration, size_t level, CFogVolumePlaneShader& fogVol);
+  static void DrawFogSlices(const zeus::CPlane* planes, size_t numPlanes, size_t iteration,
+                            const zeus::CVector3f& center, float delta, CFogVolumePlaneShader& fogVol);
   static void RenderFogVolumeModel(const zeus::CAABox& aabb, const CModel* model, const zeus::CTransform& modelMtx,
                                    const zeus::CTransform& viewMtx, const CSkinnedModel* sModel, int pass,
                                    CFogVolumePlaneShader* fvs);
